the bastion channel moves out of bootstrap
